slotswapper/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pydantic==2.9.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-multipart==0.0.12
python-dotenv==1.0.1
